requirements.txt
gradio>=4.44.0
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
openai>=1.0.0
websockets>=12.0
python-dotenv>=1.0.0
numpy>=1.24.0
sounddevice>=0.4.6
scipy>=1.11.0
soundfile>=0.12.0
pydub>=0.25.0
certifi>=2023.0.0
